refactor: move CryptographyUtils to SP.Utils project
